fix: clean up compiler warnings across all platforms
- pitch_shifter_st.cpp: cast unused receiveSamples return to (void)
- PluginGui.cpp: cast unused time_sig_denominator and dialog return to (void)
- pitch_detector_aubio.h: add missing override on set_gate and set_freq_range
- PluginProcessor.cpp: fix sign-compare using std::size_t for string loop index
- CMakeLists.txt: remove -Wno-non-virtual-dtor from global add_definitions (C++-only flag was leaking into C compilation)

+++ JUCE/Source/PluginProcessor.cpp
@@ -366,7 +366,7 @@ void AutotalentAudioProcessor::getStateInformation (MemoryBlock& destData)
     
     {
         kvbuf *array = kvbuf_create_array(&hooks);
-        for (std::int32_t i = 0; i < _misc_param.length(); i++)
+        for (std::size_t i = 0; i < _misc_param.length(); i++)
         {
             kvbuf_add_item_to_array(array, kvbuf_create_int8(&hooks, _misc_param.c_str()[i]));
         }
